nlp_service/requirements.txt
fastapi
uvicorn
google-generativeai 